# projeto-rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de controle de pacientes. Os doutores fazem login (OAuth2 password + JWT). Os dados do paciente (cadastro, consulta, alteração, remoção) ficam no serviço SOAP (projeto-soap); este REST guarda os documentos gerados em cada vinda e as finalizações de plantão. Cada vinda do paciente ao hospital é uma passagem (uma por dia) com os seus documentos: a admissão abre a passagem em aguardando_triagem, a triagem leva para aguardando_consulta e a consulta dá alta. O resumo do paciente usa só as passagens anteriores. Quadro em tempo real em GET / (WebSocket em /ws?token=...).",
        "title": "Projeto REST API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pacientes/{id}/passagens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lista todas as vindas do paciente ao hospital, com os documentos de cada uma, da mais recente para a mais antiga",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Passagens do paciente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente (local)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Passagem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Abre a passagem de hoje com a admissão e coloca o card em \"Aguardando triagem\". Se o paciente já teve alta hoje, a passagem do dia é reaberta. Responde 409 se ele já está no hospital.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Registrar chegada",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente (local)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Admissão da chegada",
                        "name": "admissao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdmissaoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Passagem do dia reaberta",
                        "schema": {
                            "$ref": "#/definitions/model.Passagem"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Passagem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pacientes/{id}/resumo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resumo para os doutores baseado só nas passagens encerradas de dias anteriores. A passagem atual nunca entra. Hoje o texto é gerado por regras; o contrato já está pronto para um gerador com IA.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Resumo do paciente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente (local)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResumoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/passagens/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uma passagem com o paciente e os documentos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Buscar passagem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da passagem",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Passagem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/passagens/{id}/admissao": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza a admissão de uma passagem aberta. Não muda a etapa.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Corrigir admissão",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da passagem",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Admissão",
                        "name": "admissao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdmissaoDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Passagem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/passagens/{id}/consulta": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Salva a consulta e encerra a passagem: o card sai do quadro e vai para \"Atendidos hoje\". Exige triagem.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Registrar consulta e dar alta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da passagem",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Consulta",
                        "name": "consulta",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConsultaDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Passagem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/quadro": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Passagens abertas (de qualquer dia) mais as de hoje que já tiveram alta. É o mesmo conteúdo que o WebSocket envia no \"init\".",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passagens"
                ],
                "summary": "Quadro",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Passagem"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AdmissaoDTO": {
            "type": "object",
            "required": [
                "forma_chegada",
                "queixa_principal"
            ],
            "properties": {
                "acompanhante": {
                    "type": "string"
                },
                "convenio": {
                    "type": "string"
                },
                "forma_chegada": {
                    "type": "string",
                    "enum": [
                        "andando",
                        "cadeira_rodas",
                        "maca",
                        "ambulancia"
                    ],
                    "example": "andando"
                },
                "observacoes": {
                    "type": "string"
                },
                "queixa_principal": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Cole o access_token obtido em POST /auth/token (com ou sem o prefixo \"Bearer \")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}